src/backend/ohlc_aggregator/requirements.txt
clickhouse-connect==0.8.15
pulsar-client==3.5.0
pydantic==2.10.5
pydantic-settings==2.7.1
